pyproject.toml
[project]
name = "capstone-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.84.0",
    "beautifulsoup4>=4.14.3",
    "chromadb>=1.5.5",
    "datasets==3.6.0",
    "feedparser>=6.0.12",
    "google-generativeai>=0.8.6",
    "gradio>=6.9.0",
    "haystack-ai>=2.25.2",
    "ipykernel>=7.2.0",
    "ipywidgets>=8.1.8",
    "jupyter-dash>=0.4.2",
    "langchain>=1.2.12",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.18",
    "langchain-openai>=1.1.11",
    "langchain-text-splitters>=1.1.1",
    "langfuse>=2.60.10",
    "llama-index>=0.10.45.post1",
    "matplotlib>=3.10.8",
    "modal>=1.3.5",
    "numpy>=2.4.3",
    "ollama>=0.6.1",
    "openai>=2.26.0",
    "opentelemetry-api>=1.40.0",
    "opentelemetry-exporter-jaeger>=1.21.0",
    "opentelemetry-sdk>=1.40.0",
    "pandas>=3.0.1",
    "plotly>=6.6.0",
    "prometheus-client>=0.24.1",
    "protobuf==3.20.2",
    "psutil>=7.2.2",
    "pydub>=0.25.1",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
    "scikit-learn>=1.8.0",
    "scipy>=1.17.1",
    "sentence-transformers>=5.3.0",
    "setuptools>=82.0.1",
    "speedtest-cli>=2.1.3",
    "torch>=2.10.0",
    "tqdm>=4.67.3",
    "transformers>=5.3.0",
    "wandb>=0.25.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["core*"]
exclude = ["tests*", "guide*", "week8*", "chroma_db*"]
